requirements.txt
scipy == any
opencv-python == any
pyyaml == any
numpy == any
matplotlib == any
